fix: do not remove scheduling configurations when adding new programs
This fixes a long standing issue which disallowed incremental channel
building via time/random slots.

Fixes #934
