--- web/src/components/slot_scheduler/commonSlotSchedulerOptions.ts
@@ -1,20 +1,0 @@
-export type DropdownOption<T extends string | number> = {
-  value: T;
-  description: string;
-};
-
-export type ProgramOption = DropdownOption<string>;
-
-export const padOptions: DropdownOption<number>[] = [
-  { value: 1, description: 'Do not pad' },
-  { value: 5 * 60 * 1000, description: '0:00, 0:05, 0:10, ..., 0:55' },
-  { value: 10 * 60 * 1000, description: '0:00, 0:10, 0:20, ..., 0:50' },
-  { value: 15 * 60 * 1000, description: '0:00, 0:15, 0:30, ..., 0:45' },
-  { value: 30 * 60 * 1000, description: '0:00, 0:30' },
-  { value: 1 * 60 * 60 * 1000, description: '0:00' },
-];
-
-export const flexOptions: DropdownOption<'end' | 'distribute'>[] = [
-  { value: 'distribute', description: 'Between videos' },
-  { value: 'end', description: 'End of the slot' },
-];
