+++ web/src/hooks/programming_controls/useSlotProgramOptions.ts
@@ -1,4 +1,4 @@
-import { ProgramOption } from '@/components/slot_scheduler/commonSlotSchedulerOptions';
+import { ProgramOption } from '@/helpers/slotSchedulerUtil';
 import { useChannelEditor } from '@/store/selectors';
 import { isUICustomProgram, isUIRedirectProgram } from '@/types';
 import { CustomShow, isContentProgram } from '@tunarr/types';
@@ -20,21 +20,28 @@ export const useSlotProgramOptions = () => {
 
   return useMemo<ProgramOption[]>(() => {
     const contentPrograms = filter(newLineup, isContentProgram);
-    const opts: ProgramOption[] = [{ value: 'flex', description: 'Flex' }];
+    const opts: ProgramOption[] = [
+      { value: 'flex', description: 'Flex', type: 'flex' },
+    ];
 
     if (contentPrograms.length) {
       if (some(contentPrograms, (p) => p.subtype === 'movie')) {
-        opts.push({ description: 'Movies', value: 'movie' });
+        opts.push({ description: 'Movies', value: 'movie', type: 'movie' });
       }
 
       const showOptions = chain(contentPrograms)
         .filter((p) => p.subtype === 'episode')
         .reject((p) => isEmpty(p.showId))
         .uniqBy((p) => p.showId)
-        .map((show) => ({
-          description: show.title,
-          value: `show.${show.showId}`,
-        }))
+        .map(
+          (show) =>
+            ({
+              description: show.title,
+              value: `show.${show.showId}`,
+              type: 'show',
+              showId: show.showId!,
+            }) satisfies ProgramOption,
+        )
         .value();
       opts.push(...showOptions);
     }
@@ -44,10 +51,15 @@ export const useSlotProgramOptions = () => {
         .filter(isUICustomProgram)
         .reject((p) => isUndefined(customShowsById[p.customShowId]))
         .uniqBy((p) => p.customShowId)
-        .map((p) => ({
-          description: customShowsById[p.customShowId].name,
-          value: `custom-show.${p.customShowId}`,
-        }))
+        .map(
+          (p) =>
+            ({
+              description: customShowsById[p.customShowId].name,
+              value: `custom-show.${p.customShowId}`,
+              id: p.customShowId,
+              type: 'custom-show',
+            }) satisfies ProgramOption,
+        )
         .value(),
     );
 
@@ -55,10 +67,15 @@ export const useSlotProgramOptions = () => {
       ...chain(newLineup)
         .filter(isUIRedirectProgram)
         .uniqBy((p) => p.channel)
-        .map((p) => ({
-          description: `Redirect to "${p.channelName}"`,
-          value: `redirect.${p.channel}`,
-        }))
+        .map(
+          (p) =>
+            ({
+              description: `Redirect to "${p.channelName}"`,
+              value: `redirect.${p.channel}`,
+              type: 'redirect',
+              channelId: p.channel,
+            }) satisfies ProgramOption,
+        )
         .value(),
     );
