+++ web/src/pages/channels/TimeSlotEditorPage.tsx
@@ -1,10 +1,20 @@
+import {
+  lineupItemAppearsInSchedule,
+  slotOptionIsScheduled,
+} from '@/helpers/slotSchedulerUtil.ts';
 import { useSlotProgramOptions } from '@/hooks/programming_controls/useSlotProgramOptions.ts';
 import { useChannelEditor } from '@/store/selectors.ts';
-import { ArrowBack, Autorenew, Delete } from '@mui/icons-material';
-import AddIcon from '@mui/icons-material/Add';
 import {
+  ArrowBack,
+  Autorenew,
+  ExpandLess,
+  ExpandMore,
+} from '@mui/icons-material';
+import {
+  Alert,
   Box,
   Button,
+  Collapse,
   Divider,
   FormControl,
   FormGroup,
@@ -12,23 +22,19 @@ import {
   Grid,
   IconButton,
   InputLabel,
+  ListItem,
   MenuItem,
   Select,
   SelectChangeEvent,
+  Stack,
   Typography,
   useMediaQuery,
   useTheme,
 } from '@mui/material';
 import { DateTimePicker } from '@mui/x-date-pickers/DateTimePicker';
-import { TimePicker } from '@mui/x-date-pickers/TimePicker';
 import { Link as RouterLink } from '@tanstack/react-router';
 import { dayjsMod, scheduleTimeSlots } from '@tunarr/shared';
-import { ChannelProgram } from '@tunarr/types';
-import {
-  TimeSlot,
-  TimeSlotProgramming,
-  TimeSlotSchedule,
-} from '@tunarr/types/api';
+import { TimeSlot, TimeSlotSchedule } from '@tunarr/types/api';
 import dayjs from 'dayjs';
 import timezone from 'dayjs/plugin/timezone';
 import utc from 'dayjs/plugin/utc';
@@ -36,34 +42,29 @@ import {
   chain,
   filter,
   first,
+  isEmpty,
   isUndefined,
   map,
-  maxBy,
   range,
   reject,
-  some,
 } from 'lodash-es';
 import { useSnackbar } from 'notistack';
 import pluralize from 'pluralize';
-import { Fragment, useCallback, useState } from 'react';
-import {
-  Control,
-  Controller,
-  UseFormSetValue,
-  useForm,
-  useWatch,
-} from 'react-hook-form';
+import { useCallback, useMemo, useState } from 'react';
+import { Controller, useForm } from 'react-hook-form';
+import { useToggle } from 'usehooks-ts';
 import Breadcrumbs from '../../components/Breadcrumbs.tsx';
 import PaddedPaper from '../../components/base/PaddedPaper.tsx';
 import ChannelProgrammingList from '../../components/channel_config/ChannelProgrammingList.tsx';
 import UnsavedNavigationAlert from '../../components/settings/UnsavedNavigationAlert.tsx';
+import { AddTimeSlotButton } from '../../components/slot_scheduler/AddTimeSlotButton.tsx';
+import { TimeSlotRow } from '../../components/slot_scheduler/TimeSlotRow.tsx';
+import { NumericFormControllerText } from '../../components/util/TypedController.tsx';
 import {
   DropdownOption,
-  ProgramOption,
   flexOptions,
   padOptions,
-} from '../../components/slot_scheduler/commonSlotSchedulerOptions.ts';
-import { NumericFormControllerText } from '../../components/util/TypedController.tsx';
+} from '../../helpers/slotSchedulerUtil.ts';
 import { zipWithIndex } from '../../helpers/util.ts';
 import { useUpdateLineup } from '../../hooks/useUpdateLineup.ts';
 import {
@@ -76,18 +77,7 @@ dayjs.extend(utc);
 dayjs.extend(timezone);
 dayjs.extend(dayjsMod);
 
-const OneDayMillis = dayjs.duration(1, 'day').asMilliseconds();
-
-// TODO Make this locale aware
-const DaysOfWeekMenuItems = [
-  { value: 0, name: 'Sunday' },
-  { value: 1, name: 'Monday' },
-  { value: 2, name: 'Tuesday' },
-  { value: 3, name: 'Wednesday' },
-  { value: 4, name: 'Thursday' },
-  { value: 5, name: 'Friday' },
-  { value: 6, name: 'Saturday' },
-];
+export const OneDayMillis = dayjs.duration(1, 'day').asMilliseconds();
 
 // mutates array
 function rotateArrayRight<T>(arr: T[], times: number): T[] {
@@ -108,7 +98,7 @@ function rotateArrayRight<T>(arr: T[], times: number): T[] {
   return arr;
 }
 
-type TimeSlotForm = Omit<TimeSlotSchedule, 'timeZoneOffset' | 'type'>;
+export type TimeSlotForm = Omit<TimeSlotSchedule, 'timeZoneOffset' | 'type'>;
 
 const latenessOptions: DropdownOption<number>[] = [
   dayjs.duration(5, 'minutes'),
@@ -140,288 +130,6 @@ const defaultTimeSlotSchedule: TimeSlotSchedule = {
   timeZoneOffset: new Date().getTimezoneOffset(),
 };
 
-const showOrderOptions = [
-  {
-    value: 'next',
-    description: 'Next Episode',
-  },
-  {
-    value: 'shuffle',
-    description: 'Shuffle',
-  },
-];
-
-const lineupItemAppearsInSchedule = (
-  slots: TimeSlot[],
-  item: ChannelProgram,
-) => {
-  return some(slots, (slot) => {
-    switch (slot.programming.type) {
-      case 'custom-show':
-        return (
-          item.type === 'custom' &&
-          item.customShowId === slot.programming.customShowId
-        );
-      case 'redirect':
-        return item.type === 'redirect';
-      case 'flex':
-        return item.type === 'flex';
-      case 'movie':
-        return (
-          (item.type === 'content' && item.subtype === 'movie') ||
-          (item.type === 'custom' && item.program?.subtype === 'movie')
-        );
-      case 'show': {
-        const showTitle = slot.programming.showId;
-        return (
-          item.type === 'content' &&
-          item.subtype === 'episode' &&
-          showTitle === (item.showId ?? item.title)
-        );
-      }
-    }
-  });
-};
-
-type AddTimeSlotButtonProps = {
-  control: Control<TimeSlotForm>;
-  setValue: UseFormSetValue<TimeSlotForm>;
-};
-
-const AddTimeSlotButton = ({ control, setValue }: AddTimeSlotButtonProps) => {
-  const currentSlots = useWatch({ control, name: 'slots' });
-
-  const addSlot = useCallback(() => {
-    const maxSlot = maxBy(currentSlots, (p) => p.startTime);
-    const newStartTime = maxSlot
-      ? dayjs.duration(maxSlot.startTime).add(1, 'hour')
-      : dayjs.duration(0);
-
-    const newSlots: TimeSlot[] = [
-      ...currentSlots,
-      {
-        programming: { type: 'flex' },
-        startTime: newStartTime.asMilliseconds(),
-        order: 'next',
-      },
-    ];
-
-    setValue('slots', newSlots, { shouldDirty: true });
-  }, [currentSlots, setValue]);
-
-  return (
-    <Button
-      startIcon={<AddIcon />}
-      variant="contained"
-      onClick={() => addSlot()}
-    >
-      Add Slot
-    </Button>
-  );
-};
-
-type TimeSlotProps = {
-  slot: TimeSlot;
-  index: number;
-  control: Control<TimeSlotForm>;
-  setValue: UseFormSetValue<TimeSlotForm>;
-  programOptions: ProgramOption[];
-};
-
-const TimeSlotRow = ({
-  slot,
-  index,
-  control,
-  setValue,
-  programOptions,
-}: TimeSlotProps) => {
-  const start = dayjs.tz().startOf('day');
-  const currentSlots = useWatch({ control, name: 'slots' });
-  const currentPeriod = useWatch({ control, name: 'period' });
-
-  const updateSlotTime = useCallback(
-    (idx: number, time: dayjs.Dayjs) => {
-      setValue(
-        `slots.${idx}.startTime`,
-        time
-          .mod(dayjs.duration(1, 'day'))
-          .subtract({ minutes: new Date().getTimezoneOffset() })
-          .asMilliseconds(),
-        { shouldDirty: true },
-      );
-    },
-    [setValue],
-  );
-
-  const removeSlot = useCallback(
-    (idx: number) => {
-      setValue(
-        'slots',
-        reject(currentSlots, (_, i) => idx === i),
-        { shouldDirty: true },
-      );
-    },
-    [currentSlots, setValue],
-  );
-
-  const updateSlotDay = useCallback(
-    (idx: number, currentDay: number, dayOfWeek: number) => {
-      const slot = currentSlots[idx];
-      const daylessStartTime = slot.startTime - currentDay * OneDayMillis;
-      const newStartTime = daylessStartTime + dayOfWeek * OneDayMillis;
-      setValue(`slots.${idx}.startTime`, newStartTime, { shouldDirty: true });
-    },
-    [currentSlots, setValue],
-  );
-
-  const updateSlotType = useCallback(
-    (idx: number, slotId: string) => {
-      let slotProgram: TimeSlotProgramming;
-
-      if (slotId.startsWith('show')) {
-        slotProgram = {
-          type: 'show',
-          showId: slotId.split('.')[1],
-        };
-      } else if (slotId.startsWith('movie')) {
-        slotProgram = {
-          type: 'movie',
-        };
-      } else if (slotId.startsWith('flex')) {
-        slotProgram = {
-          type: 'flex',
-        };
-      } else if (slotId.startsWith('redirect')) {
-        slotProgram = {
-          type: 'redirect',
-          channelId: slotId.split('.')[1],
-        };
-      } else if (slotId.startsWith('custom-show')) {
-        slotProgram = {
-          type: 'custom-show',
-          customShowId: slotId.split('.')[1],
-        };
-      } else {
-        return;
-      }
-
-      const slot: Omit<TimeSlot, 'startTime'> = {
-        order: 'next',
-        programming: slotProgram,
-      };
-
-      const curr = currentSlots[idx];
-
-      setValue(
-        `slots.${idx}`,
-        { ...slot, startTime: curr.startTime },
-        { shouldDirty: true },
-      );
-    },
-    [currentSlots, setValue],
-  );
-
-  const startTime = start.add(slot.startTime);
-  // .subtract(new Date().getTimezoneOffset(), 'minutes');
-  let selectValue: string;
-  switch (slot.programming.type) {
-    case 'show': {
-      selectValue = `show.${slot.programming.showId}`;
-      break;
-    }
-    case 'redirect': {
-      selectValue = `redirect.${slot.programming.channelId}`;
-      break;
-    }
-    case 'custom-show': {
-      selectValue = `${slot.programming.type}.${slot.programming.customShowId}`;
-      break;
-    }
-    default: {
-      selectValue = slot.programming.type;
-      break;
-    }
-  }
-
-  const isShowType = slot.programming.type === 'show';
-  let showInputSize = currentPeriod === 'week' ? 7 : 9;
-  if (isShowType) {
-    showInputSize -= 3;
-  }
-
-  const dayOfTheWeek = Math.floor(slot.startTime / OneDayMillis);
-
-  return (
-    <Fragment key={`${slot.startTime}_${index}`}>
-      {currentPeriod === 'week' ? (
-        <Grid item xs={2}>
-          <Select
-            fullWidth
-            value={dayOfTheWeek}
-            onChange={(e) =>
-              updateSlotDay(index, dayOfTheWeek, e.target.value as number)
-            }
-          >
-            {map(DaysOfWeekMenuItems, ({ value, name }) => (
-              <MenuItem key={value} value={value}>
-                {name}
-              </MenuItem>
-            ))}
-          </Select>
-        </Grid>
-      ) : null}
-      <Grid item xs={2}>
-        <TimePicker
-          onChange={(value) => value && updateSlotTime(index, value)}
-          value={startTime}
-          label="Start Time"
-        />
-      </Grid>
-      <Grid item xs={showInputSize}>
-        <FormControl fullWidth>
-          <InputLabel>Program</InputLabel>
-          <Select
-            label="Program"
-            value={selectValue}
-            onChange={(e) => updateSlotType(index, e.target.value)}
-          >
-            {map(programOptions, ({ description, value }) => (
-              <MenuItem key={value} value={value}>
-                {description}
-              </MenuItem>
-            ))}
-          </Select>
-        </FormControl>
-      </Grid>
-      {isShowType && (
-        <Grid item xs={3}>
-          <FormControl fullWidth>
-            <InputLabel>Order</InputLabel>
-            <Controller
-              control={control}
-              name={`slots.${index}.order`}
-              render={({ field }) => (
-                <Select label="Order" {...field}>
-                  {map(showOrderOptions, ({ description, value }) => (
-                    <MenuItem key={value} value={value}>
-                      {description}
-                    </MenuItem>
-                  ))}
-                </Select>
-              )}
-            />
-          </FormControl>
-        </Grid>
-      )}
-      <Grid item xs={1}>
-        <IconButton onClick={() => removeSlot(index)} color="error">
-          <Delete />
-        </IconButton>
-      </Grid>
-    </Fragment>
-  );
-};
-
 export default function TimeSlotEditorPage() {
   // Requires that the channel was already loaded... not the case if
   // we navigated directly, so we need to handle that
@@ -439,6 +147,7 @@ export default function TimeSlotEditorPage() {
   const theme = useTheme();
   const smallViewport = useMediaQuery(theme.breakpoints.down('sm'));
   const programOptions = useSlotProgramOptions();
+  const [unscheduledOpen, toggleUnscheduledOpen] = useToggle(false);
 
   const {
     control,
@@ -466,6 +175,14 @@ export default function TimeSlotEditorPage() {
   // Have to use a watch here because rendering depends on this value
   const currentPeriod = watch('period');
   const currentSlots = watch('slots');
+
+  const unscheduledOptions = useMemo(
+    () =>
+      reject(programOptions, (item) =>
+        slotOptionIsScheduled(currentSlots, item),
+      ),
+    [currentSlots, programOptions],
+  );
 
   const [generatedList, setGeneratedList] = useState<
     UIChannelProgram[] | undefined
@@ -633,204 +350,242 @@ export default function TimeSlotEditorPage() {
   return (
     <div>
       <Breadcrumbs />
-      <Typography variant="h4" sx={{ mb: 2 }}>
-        Edit Time Slots (Channel {channel?.number})
-      </Typography>
-      <PaddedPaper sx={{ mb: 2 }}>
-        <Typography sx={{ flexGrow: 1, fontWeight: 600 }}>
-          Time Slots
+      <Stack gap={2} useFlexGap>
+        <Typography variant="h4">
+          Edit Time Slots (Channel {channel?.number})
         </Typography>
-        <Divider sx={{ my: 2 }} />
-        {renderTimeSlots()}
-        <AddTimeSlotButton control={control} setValue={setValue} />
-        <Divider sx={{ my: 2 }} />
-        <Typography sx={{ flexGrow: 1, fontWeight: '600' }}>
-          Settings
-        </Typography>
-        <Box>
-          <Grid
-            container
-            spacing={2}
-            columns={16}
-            justifyContent={'flex-start'}
+        {!isEmpty(unscheduledOptions) && (
+          <Alert
+            severity="warning"
+            action={
+              <IconButton
+                aria-label="close"
+                color="inherit"
+                size="small"
+                onClick={() => {
+                  toggleUnscheduledOpen();
+                }}
+              >
+                {!unscheduledOpen ? (
+                  <ExpandMore fontSize="inherit" />
+                ) : (
+                  <ExpandLess fontSize="inherit" />
+                )}
+              </IconButton>
+            }
           >
-            <Grid item sm={16} md={5}>
-              <FormControl fullWidth margin="normal">
-                <InputLabel>Period</InputLabel>
-                <Controller
-                  control={control}
-                  name="period"
-                  render={({ field }) => (
-                    <Select
-                      label="Period"
-                      {...field}
-                      onChange={(e) => handlePeriodChange(e)}
-                    >
-                      <MenuItem value="day">Daily</MenuItem>
-                      <MenuItem value="week">Weekly</MenuItem>
-                    </Select>
-                  )}
-                />
-                <FormHelperText>
-                  By default, time slots are time of the day-based, you can
-                  change it to time of the day + day of the week. That means
-                  scheduling 7x the number of time slots. If you change from
-                  daily to weekly, the current schedule will be repeated 7
-                  times. If you change from weekly to daily, many of the slots
-                  will be deleted.
-                </FormHelperText>
-              </FormControl>
-            </Grid>
-            <Grid item sm={16} md={5}>
-              <FormControl fullWidth margin="normal">
-                <InputLabel>Max Lateness</InputLabel>
-                <Controller
-                  control={control}
-                  name="latenessMs"
-                  render={({ field }) => (
-                    <Select label="Max Lateness" {...field}>
-                      {latenessOptions.map((opt) => (
-                        <MenuItem key={opt.value} value={opt.value}>
-                          {opt.description}
-                        </MenuItem>
-                      ))}
-                    </Select>
-                  )}
-                />
-
-                <FormHelperText>
-                  Allows programs to play a bit late if the previous program
-                  took longer than usual. If a program is too late, Flex is
-                  scheduled instead.
-                </FormHelperText>
-              </FormControl>
-            </Grid>
-            <Grid item sm={16} md={5}>
-              <FormControl fullWidth margin="normal">
-                <InputLabel>Pad Times</InputLabel>
-                <Controller
-                  control={control}
-                  name="padMs"
-                  render={({ field }) => (
-                    <Select label="Pad Times" {...field}>
-                      {padOptions.map((opt) => (
-                        <MenuItem key={opt.value} value={opt.value}>
-                          {opt.description}
-                        </MenuItem>
-                      ))}
-                    </Select>
-                  )}
-                />
-
-                <FormHelperText>
-                  Ensures programs have a nice-looking start time, it will add
-                  Flex time to fill the gaps.
-                </FormHelperText>
-              </FormControl>
-            </Grid>
-
-            <Grid item sm={16} md={5}>
-              <FormControl fullWidth margin="normal">
-                <InputLabel>Flex Style</InputLabel>
-                <Controller
-                  control={control}
-                  name="flexPreference"
-                  render={({ field }) => (
-                    <Select label="Flex Style" {...field}>
-                      {flexOptions.map((opt) => (
-                        <MenuItem key={opt.value} value={opt.value}>
-                          {opt.description}
-                        </MenuItem>
-                      ))}
-                    </Select>
-                  )}
-                />
-                <FormHelperText>
-                  Usually slots need to add flex time to ensure that the next
-                  slot starts at the correct time. When there are multiple
-                  videos in the slot, you might prefer to distribute the flex
-                  time between the videos or to place most of the flex time at
-                  the end of the slot.
-                </FormHelperText>
-              </FormControl>
-            </Grid>
-            <Grid item sm={16} md={5}>
-              <FormGroup row>
-                <NumericFormControllerText
-                  control={control}
-                  name="maxDays"
-                  prettyFieldName="Days to Precalculate"
-                  TextFieldProps={{
-                    label: 'Days to Precalculate',
-                    fullWidth: true,
-                    margin: 'normal',
-                  }}
-                />
-
-                <FormHelperText sx={{ ml: 1 }}>
-                  Maximum number of days to precalculate the schedule. Note that
-                  the length of the schedule is also bounded by the maximum
-                  number of programs allowed in a channel.
-                  <br />
-                  <strong>
-                    Note: Previewing the schedule in the browser for long
-                    lengths of time can cause UI performance issues
-                  </strong>
-                </FormHelperText>
-              </FormGroup>
-            </Grid>
-          </Grid>
-        </Box>
-        <Divider sx={{ my: 4 }} />
-        <Box sx={{ display: 'flex', justifyContent: 'center', my: 4 }}>
-          <Button
-            variant="contained"
-            onClick={() => calculateSlots()}
-            disabled={!isValid || !isDirty}
-            startIcon={<Autorenew />}
-          >
-            Refresh Preview
-          </Button>
-        </Box>
-      </PaddedPaper>
-      <PaddedPaper>
-        <Typography sx={{ mb: 1 }}>Programming Preview</Typography>
-        <Divider sx={{ mb: 1 }} />
-        <DateTimePicker
-          label="Programming Start"
-          value={startTime}
-          disabled
-          slotProps={{ textField: { size: 'small' } }}
-        />
-        {generatedList ? (
-          <ChannelProgrammingList
-            type={'direct'}
-            programList={zipWithIndex(generatedList)}
-            enableDnd={false}
-            enableRowDelete={false}
-            enableRowEdit={false}
-            virtualListProps={{
-              width: '100%',
-              height: 400,
-              itemSize: smallViewport ? 70 : 35,
-              overscanCount: 5,
-            }}
-          />
-        ) : (
-          <ChannelProgrammingList
-            type="selector"
-            enableDnd={false}
-            enableRowDelete={false}
-            enableRowEdit={false}
-            virtualListProps={{
-              width: '100%',
-              height: 400,
-              itemSize: smallViewport ? 70 : 35,
-              overscanCount: 5,
-            }}
-          />
+            There are {unscheduledOptions.length} unscheduled{' '}
+            {pluralize('program', unscheduledOptions.length)}. Unscheduled items
+            will be removed from the channel when saving.
+            <Collapse in={unscheduledOpen}>
+              <>
+                {map(unscheduledOptions, (option) => (
+                  <ListItem key={option.value}>
+                    {option.description} ({option.type})
+                  </ListItem>
+                ))}
+              </>
+            </Collapse>
+          </Alert>
         )}
-      </PaddedPaper>
+        <PaddedPaper>
+          <Stack direction="row" alignItems="center">
+            <Typography sx={{ flexGrow: 1, fontWeight: 600 }}>
+              Time Slots
+            </Typography>
+            <AddTimeSlotButton control={control} setValue={setValue} />
+          </Stack>
+          <Divider sx={{ my: 2 }} />
+          {renderTimeSlots()}
+          <Divider sx={{ my: 2 }} />
+          <Typography sx={{ flexGrow: 1, fontWeight: '600' }}>
+            Settings
+          </Typography>
+          <Box>
+            <Grid
+              container
+              spacing={2}
+              columns={16}
+              justifyContent={'flex-start'}
+            >
+              <Grid item sm={16} md={5}>
+                <FormControl fullWidth margin="normal">
+                  <InputLabel>Period</InputLabel>
+                  <Controller
+                    control={control}
+                    name="period"
+                    render={({ field }) => (
+                      <Select
+                        label="Period"
+                        {...field}
+                        onChange={(e) => handlePeriodChange(e)}
+                      >
+                        <MenuItem value="day">Daily</MenuItem>
+                        <MenuItem value="week">Weekly</MenuItem>
+                      </Select>
+                    )}
+                  />
+                  <FormHelperText>
+                    By default, time slots are time of the day-based, you can
+                    change it to time of the day + day of the week. That means
+                    scheduling 7x the number of time slots. If you change from
+                    daily to weekly, the current schedule will be repeated 7
+                    times. If you change from weekly to daily, many of the slots
+                    will be deleted.
+                  </FormHelperText>
+                </FormControl>
+              </Grid>
+              <Grid item sm={16} md={5}>
+                <FormControl fullWidth margin="normal">
+                  <InputLabel>Max Lateness</InputLabel>
+                  <Controller
+                    control={control}
+                    name="latenessMs"
+                    render={({ field }) => (
+                      <Select label="Max Lateness" {...field}>
+                        {latenessOptions.map((opt) => (
+                          <MenuItem key={opt.value} value={opt.value}>
+                            {opt.description}
+                          </MenuItem>
+                        ))}
+                      </Select>
+                    )}
+                  />
+
+                  <FormHelperText>
+                    Allows programs to play a bit late if the previous program
+                    took longer than usual. If a program is too late, Flex is
+                    scheduled instead.
+                  </FormHelperText>
+                </FormControl>
+              </Grid>
+              <Grid item sm={16} md={5}>
+                <FormControl fullWidth margin="normal">
+                  <InputLabel>Pad Times</InputLabel>
+                  <Controller
+                    control={control}
+                    name="padMs"
+                    render={({ field }) => (
+                      <Select label="Pad Times" {...field}>
+                        {padOptions.map((opt) => (
+                          <MenuItem key={opt.value} value={opt.value}>
+                            {opt.description}
+                          </MenuItem>
+                        ))}
+                      </Select>
+                    )}
+                  />
+
+                  <FormHelperText>
+                    Ensures programs have a nice-looking start time, it will add
+                    Flex time to fill the gaps.
+                  </FormHelperText>
+                </FormControl>
+              </Grid>
+
+              <Grid item sm={16} md={5}>
+                <FormControl fullWidth margin="normal">
+                  <InputLabel>Flex Style</InputLabel>
+                  <Controller
+                    control={control}
+                    name="flexPreference"
+                    render={({ field }) => (
+                      <Select label="Flex Style" {...field}>
+                        {flexOptions.map((opt) => (
+                          <MenuItem key={opt.value} value={opt.value}>
+                            {opt.description}
+                          </MenuItem>
+                        ))}
+                      </Select>
+                    )}
+                  />
+                  <FormHelperText>
+                    Usually slots need to add flex time to ensure that the next
+                    slot starts at the correct time. When there are multiple
+                    videos in the slot, you might prefer to distribute the flex
+                    time between the videos or to place most of the flex time at
+                    the end of the slot.
+                  </FormHelperText>
+                </FormControl>
+              </Grid>
+              <Grid item sm={16} md={5}>
+                <FormGroup row>
+                  <NumericFormControllerText
+                    control={control}
+                    name="maxDays"
+                    prettyFieldName="Days to Precalculate"
+                    TextFieldProps={{
+                      label: 'Days to Precalculate',
+                      fullWidth: true,
+                      margin: 'normal',
+                    }}
+                  />
+
+                  <FormHelperText sx={{ ml: 1 }}>
+                    Maximum number of days to precalculate the schedule. Note
+                    that the length of the schedule is also bounded by the
+                    maximum number of programs allowed in a channel.
+                    <br />
+                    <strong>
+                      Note: Previewing the schedule in the browser for long
+                      lengths of time can cause UI performance issues
+                    </strong>
+                  </FormHelperText>
+                </FormGroup>
+              </Grid>
+            </Grid>
+          </Box>
+          <Divider sx={{ my: 4 }} />
+          <Box sx={{ display: 'flex', justifyContent: 'center', my: 4 }}>
+            <Button
+              variant="contained"
+              onClick={() => calculateSlots()}
+              disabled={!isValid || !isDirty}
+              startIcon={<Autorenew />}
+            >
+              Refresh Preview
+            </Button>
+          </Box>
+        </PaddedPaper>
+        <PaddedPaper>
+          <Typography sx={{ mb: 1 }}>Programming Preview</Typography>
+          <Divider sx={{ mb: 1 }} />
+          <DateTimePicker
+            label="Programming Start"
+            value={startTime}
+            disabled
+            slotProps={{ textField: { size: 'small' } }}
+          />
+          {generatedList ? (
+            <ChannelProgrammingList
+              type={'direct'}
+              programList={zipWithIndex(generatedList)}
+              enableDnd={false}
+              enableRowDelete={false}
+              enableRowEdit={false}
+              virtualListProps={{
+                width: '100%',
+                height: 400,
+                itemSize: smallViewport ? 70 : 35,
+                overscanCount: 5,
+              }}
+            />
+          ) : (
+            <ChannelProgrammingList
+              type="selector"
+              enableDnd={false}
+              enableRowDelete={false}
+              enableRowEdit={false}
+              virtualListProps={{
+                width: '100%',
+                height: 400,
+                itemSize: smallViewport ? 70 : 35,
+                overscanCount: 5,
+              }}
+            />
+          )}
+        </PaddedPaper>
+      </Stack>
       <UnsavedNavigationAlert isDirty={isDirty} />
       <Box sx={{ display: 'flex', justifyContent: 'end', pt: 1, columnGap: 1 }}>
         <Box flexGrow={1}>
